Reset the delegations modal
Because the XHR request of the vote button renders
`update_vote_button.js.erb` the vote_button partial was re-rendered and
so the delegations modal but action_delegator.js wasn't re-evaluated. As
a result, the click event listener was gone from the delegation link
making it impossible to open the delegations modal.

We followed the same approach of `update_vote_button.js.erb` and reset
the modal and the button. The Decidim's regular vote button doesn't
re-attach the click handler because the revoke button is a <button> tag
that submits a form through XHR so it doesn't need any event listener.